Move operation parser to core
