+++ dap2-service/src/test/scala/latis/util/dap2/parser/TestConstraintParser.scala
@@ -3,6 +3,7 @@ package latis.util.dap2.parser
 import org.junit._, Assert._
 import org.scalatestplus.junit.JUnitSuite
 
+import latis.ops.parser.ast._
 import ast._
 
 class TestConstraintParser extends JUnitSuite {
